+++ src/pki_engine/VaultPKIEngine.js
@@ -37,6 +37,7 @@ class VaultPKIEngine extends PKIEngine {
     pkiServerRole,
     pkiClientRole,
     signExpiryHours,
+    bundleTtl,
     keyLength,
     keyAlgorithm
   }) {
@@ -48,6 +49,7 @@ class VaultPKIEngine extends PKIEngine {
     this.pkiClientRole = pkiClientRole;
     this.mounts = mounts;
     this.signExpiryHours = signExpiryHours;
+    this.bundleTtl = bundleTtl;
     this.reconnectTimer = null;
     this.keyLength = keyLength;
     this.keyAlgorithm = keyAlgorithm;
@@ -307,6 +309,8 @@ class VaultPKIEngine extends PKIEngine {
       .sort((a, b) => b.id - a.id)[0];
     const cert = this.getCertInfo(dfspClientCert.certificate);
     const bundle = {
+      timeStamp: new Date().toISOString().substring(0,19).replace(/T|:/g,'-'),
+      ttl: this.bundleTtl,
       ca_bundle: `${dfspCA.intermediateChain}\n${dfspCA.rootCertificate}`,
       client_key: dfspClientCert.key,
       client_cert_chain: `${dfspClientCert.certificate}\n${dfspCA.intermediateChain}\n${dfspCA.rootCertificate}`,
@@ -317,6 +321,7 @@ class VaultPKIEngine extends PKIEngine {
       isProxy,
     };
     await this.client.write(`${this.mounts.dfspClientCertBundle}/${dfspName}`, bundle);
+    return bundle;
   }
 
   async populateDFSPInternalIPWhitelistBundle (value) {